WIP: Events on pub/sub event bus should be in the commons package.

+++ internal/channels/batch_open_channel.go
@@ -89,7 +89,7 @@ func checkPrepareReq(bocReq commons.BatchOpenRequest) (req *lnrpc.BatchOpenChann
 		}
 
 		if channel.MinHtlcMsat != nil {
-			boChannel.MinHtlcMsat = *channel.MinHtlcMsat
+			boChannel.MinHtlcMsat = int64(*channel.MinHtlcMsat)
 		}
 
 		if channel.PushSat != nil {
